refactor: 삭제 핸들러 커스텀 훅
